Fixing an accidental conversion from double to int
git-svn-id: https://svn.divms.uiowa.edu/repos/clc/projects/starexec/dev/branches/fb1/starexec@19595 bd178f5f-7bea-4014-844f-495ca19b8805

+++ src/org/starexec/data/database/Jobs.java
@@ -1696,8 +1696,8 @@ public class Jobs {
 				
 				JobPair jp=idsToPairs.get(results.getInt("job_pairs.id"));
 				JoblineStage stage=new JoblineStage();
-				stage.setCpuUsage(results.getInt("cpu"));
-				stage.setWallclockTime(results.getInt("wallclock"));
+				stage.setCpuUsage(results.getDouble("cpu"));
+				stage.setWallclockTime(results.getDouble("wallclock"));
 				stage.setStageId(results.getInt("pipeline_stages.stage_id"));
 				//everything below this line is in a stage
 				id=results.getInt("solver_id");
